Reorganize files for module aliasing (#5529)
Resolves rdar://93611579
